backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
gunicorn>=21.2
whitenoise>=6.6
